rename module to ins v2
