+++ src/smsfusion/__init__.py
@@ -3,7 +3,7 @@ from ._coning_sculling import ConingScullingAlg
 from ._ins import AHRS, VRU, AidedINS, FixedNED, StrapdownINS, gravity
 from ._smoothing import FixedIntervalSmoother
 from ._transforms import quaternion_from_euler
-from ._v2 import AHRSv2
+from ._ins_v2 import AHRSv2
 
 __all__ = [
     "AHRS",
